# Faster prodigal.