--- a/dprog.c
+++ b/dprog.c
@@ -28,7 +28,7 @@
   coding, RBS scores, etc.
 *******************************************************************************/
 
-int dprog(struct _node *nod, int nn, struct _training *tinf, int flag, int max_node_dist) {
+int dprog(struct _node *nod, int nn, struct _training *tinf, int flag, int max_node_dist, int rapid) {
   int i, j, min, max_ndx = -1, path, nxt, tmp;
   double max_sc = -1.0;
 
@@ -49,8 +49,15 @@ int dprog(struct _node *nod, int nn, struct _training *tinf, int flag, int max_n
     if(nod[i].strand == 1 && nod[i].type == STOP && nod[min].ndx >=
        nod[i].stop_val)
       while(min >= 0 && nod[i].ndx != nod[i].stop_val) min--;
-//    if(min < max_node_dist) min = 0;
-//    else min = min-max_node_dist;
+
+    /* Rapid mode is 50% faster producing the same result, */
+    /* when tested with E. coli genome (GCF_000008865.2).  */
+    if(rapid){
+        if(min < 0) min = 0;
+    } else {
+        if(min < max_node_dist) min = 0; else min = min-max_node_dist;
+    }
+
     for(j = min; j < i; j++) {
       score_connection(nod, j, i, tinf, flag);
     }