--- a/dprog.h
+++ b/dprog.h
@@ -30,7 +30,7 @@
 #define MAX_OPP_OVLP 200
 // #define MAX_NODE_DIST 500
 
-int dprog(struct _node *, int, struct _training *, int, int);
+int dprog(struct _node *, int, struct _training *, int, int, int);
 void score_connection(struct _node *, int, int, struct _training *, int);
 void eliminate_bad_genes(struct _node *, int, struct _training *);